# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https:///www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://kids-collage.test/api",
            "description": "Local Host"
        }
    ],
    "paths": {
...
        "/blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "get blogs list",
                "description": "Returns list of blogs",
                "operationId": "getBlogs",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/page_limit"
                    },
                    {
                        "$ref": "#/components/parameters/search"
                    },
                    {
                        "$ref": "#/components/parameters/advanced_search"
                    },
                    {
                        "$ref": "#/components/parameters/sort"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BlogResource"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "default": "http://localhost/api/blog?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "default": "http://localhost/api/blog?page=4"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "default": "null",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "default": "http://localhost/api/blog?page=2",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Meta"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "No message"
                                        },
                                        "advance_search_field": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "key": {
                                                        "type": "string",
                                                        "default": "id"
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "default": "text"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "default": "number"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "extra": {
                                            "properties": {
                                                "default_sort": {
                                                    "type": "string",
                                                    "default": "-id"
                                                },
                                                "sorts": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "default": [
                                                        "id",
                                                        "created_at",
                                                        "updated_at"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Store new blog",
                "description": "Returns new blog data",
                "operationId": "storeBlog",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBlogRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBlogRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "blog has been stored successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BlogResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/{blog}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get blog information",
                "description": "Returns blog data",
                "operationId": "getBlogByID",
                "parameters": [
                    {
                        "name": "blog",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "No message"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BlogDetailResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Blog"
                ],
                "summary": "Update existing blog",
                "description": "Returns updated blog data",
                "operationId": "updateBlog",
                "parameters": [
                    {
                        "name": "blog",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBlogRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBlogRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "blog has been updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BlogResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Blog"
                ],
                "summary": "Delete existing blog",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteBlog",
                "parameters": [
                    {
                        "name": "blog",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "blog has been deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/toggle/{blog}": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Toggle Blog",
                "operationId": "toggleBlog",
                "parameters": [
                    {
                        "name": "blog",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "blog has been toggled successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BlogResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Meta": {
                "title": "Pagination",
                "properties": {
                    "current_page": {
                        "description": "Current page number",
                        "type": "integer",
                        "default": 1
                    },
                    "from": {
                        "description": "First page number",
                        "type": "integer",
                        "default": 1
                    },
                    "last_page": {
                        "description": "Last page number",
                        "type": "integer",
                        "default": 3
                    },
                    "links": {
                        "description": "Links of pages",
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "description": "Url of page",
                                    "type": "string",
                                    "default": "http://localhost:8000/api/shift?page=1"
                                },
                                "label": {
                                    "description": "Label of page",
                                    "type": "string",
                                    "default": "1"
                                },
                                "active": {
                                    "description": "this page is active or not",
                                    "type": "boolean",
                                    "default": true
                                }
                            },
                            "type": "object"
                        },
                        "default": [
                            {
                                "url": null,
                                "label": "&laquo; Previous",
                                "active": false
                            },
                            {
                                "url": "http://localhost:8000/api/category?page=1",
                                "label": "1",
                                "active": true
                            },
                            {
                                "url": "http://localhost:8000/api/category?page=2",
                                "label": "2",
                                "active": false
                            },
                            {
                                "url": "http://127.0.0.1:8000/api/category?page=3",
                                "label": "3",
                                "active": false
                            },
                            {
                                "url": "http://localhost:8000/api/category?page=2",
                                "label": "Next &raquo;",
                                "active": false
                            }
                        ]
                    },
                    "path": {
                        "description": "Base url path",
                        "type": "string",
                        "default": "http://localhost:8000/api/category"
                    },
                    "per_page": {
                        "description": "Number of items per page",
                        "type": "integer",
                        "default": 15
                    },
                    "to": {
                        "description": "Last page number",
                        "type": "integer",
                        "default": 10
                    },
                    "total": {
                        "description": "number of records",
                        "type": "integer",
                        "default": 50
                    }
                },
                "type": "object"
            },
...
            "BlogResource": {
                "title": "BlogResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "default": "1"
                    },
                    "title": {
                        "type": "string",
                        "default": "Blog Title"
                    },
                    "description": {
                        "type": "string",
                        "default": "Blog Description"
                    },
                    "slug": {
                        "type": "string",
                        "default": "blog-title"
                    },
                    "published": {
                        "type": "boolean",
                        "default": true
                    },
                    "user": {
                        "$ref": "#/components/schemas/SimpleUserResource"
                    },
                    "category": {
                        "type": "object"
                    },
                    "view_count": {
                        "type": "integer",
                        "default": "100"
                    },
                    "comment_count": {
                        "type": "integer",
                        "default": "10"
                    },
                    "wish_count": {
                        "type": "integer",
                        "default": "50"
                    },
                    "updated_at": {
                        "type": "string",
                        "default": "2024-08-19T07:26:07.000000Z"
                    },
                    "created_at": {
                        "type": "string",
                        "default": "2024-08-19T07:26:07.000000Z"
                    }
                },
                "type": "object"
            },
            "SimpleUserResource": {
                "title": "SimpleUserResource",
                "properties": {
                    "id": {
                        "description": "User id",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "John",
                        "nullable": true
                    },
                    "family": {
                        "type": "string",
                        "example": "Doe",
                        "nullable": true
                    },
                    "mobile": {
                        "type": "string",
                        "example": "09379151111"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
            "advanced_search": {
                "name": "filter[a_search][]",
                "in": "query",
                "description": "Search in events",
                "required": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "column": {
                                "type": "string",
                                "default": "id"
                            },
                            "operator": {
                                "type": "string",
                                "default": "=",
                                "enum": [
                                    "like",
                                    "=",
                                    ">",
                                    "<",
                                    ">=",
                                    "<="
                                ]
                            },
                            "from": {
                                "type": "string",
                                "default": "1"
                            },
                            "to": {
                                "type": "string",
                                "default": "",
                                "nullable": true
                            },
                            "contain": {
                                "type": "integer",
                                "default": "1",
                                "enum": [
                                    0,
                                    1
                                ]
                            }
                        },
                        "type": "object"
                    }
                }
            },
...
            "search": {
                "name": "filter[search]",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": ""
                }
            },
...
            "sort": {
                "name": "sort",
                "in": "query",
                "description": "Sort criteria for shifts (e.g., name, -created_at)",
                "required": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "default": [
                        "-id"
                    ]
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "page number",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "page_limit": {
                "name": "page_limit",
                "in": "query",
                "description": "number of items per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 15
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header"
            },
            "clubId": {
                "type": "apiKey",
                "description": "Club ID header",
                "name": "X-Club-Id",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": [],
            "clubId": []
        }
    ]
}